middleware reorganization completed
