+++ server/trpc.ts
@@ -3,11 +3,6 @@ import superjson from 'superjson';
 import { Prisma } from '@prisma/client';
 import { ZodError } from 'zod';
 import { Context } from './context';
-import z from 'zod';
-import { TRPCError } from '@trpc/server';
-import { generateTypedError } from './errorTypes';
-import prisma from '@/prisma';
-import { zCourseRoles, zSiteRoles } from '@/types/sharedZodTypes';
 
 export const trpc = initTRPC.context<Context>().create({
   transformer: superjson,
@@ -29,77 +24,5 @@ export const trpc = initTRPC.context<Context>().create({
   }
 });
 
-const courseMemberInput = z.object({
-  courseMemberId: z.string()
-});
-
-
-const isElevatedCourseMemberForCourseMember = trpc.middleware(
-  async ({ next, ctx, rawInput }) => {
-    const email = ctx.session?.email;
-
-    const result = courseMemberInput.safeParse(rawInput);
-
-    if (!email)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'UNAUTHORIZED',
-          message: 'TRPC Middleware: User does not have a valid JWT'
-        })
-      );
-
-    if (!result.success)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'BAD_REQUEST',
-          message:
-            'TRPC Middleware: isElevatedCourseMemberForCourseMember requires a valid courseMemberId'
-        })
-      );
-
-    const ctxCourseMembership = await prisma.courseMember.findFirst({
-      where: {
-        id: result.data.courseMemberId
-      }
-    });
-
-    if (!ctxCourseMembership)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'BAD_REQUEST',
-          message: 'TRPC Middleware: Could not find courseMember with id.'
-        })
-      );
-
-    //Find the first courseMember who is either a professor or TA
-    const courseMember = await prisma.courseMember.findFirst({
-      where: {
-        courseId: ctxCourseMembership.courseId,
-        email: email,
-        OR: [
-          { role: zCourseRoles.enum.teacher },
-          { role: zCourseRoles.enum.ta }
-        ]
-      }
-    });
-
-    if (!courseMember)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'UNAUTHORIZED',
-          message:
-            'TRPC Middleware: User either does not exist in course or does not have elevated priveleges'
-        })
-      );
-
-    return next();
-  }
-);
-
 export const router = trpc.router;
 export const publicProcedure = trpc.procedure;
-
-/* -------- Checks privileges using courseId -------- */
-export const elevatedCourseMemberForCourseMemberProcedure = trpc.procedure.use(
-  isElevatedCourseMemberForCourseMember
-);
